requirements.txt
alembic~=1.13.1
psycopg2-binary==2.9.9
SQLAlchemy~=2.0.25
fastapi~=0.109.0
uvicorn
python-telegram-bot
python-dotenv~=1.0.0